+++ tests/testthat.R
@@ -1,5 +1,4 @@
 library(testthat)
 library(EGM)
-EGM::set_wfdb_path("/usr/local/bin")
 
 test_check("EGM")